+++ src/main/java/woowacourse/shoppingcart/dao/CustomerDao.java
@@ -12,7 +12,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.domain.customer.Customer;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 
 @Repository
@@ -82,7 +82,8 @@ public class CustomerDao {
 
     public boolean update(Customer customer) {
         final String query = "UPDATE customer SET address = ?, phone_number = ? WHERE id = ?";
-        return isUpdated(jdbcTemplate.update(query, customer.getAddress(), customer.getPhoneNumber(), customer.getId()));
+        return isUpdated(
+                jdbcTemplate.update(query, customer.getAddress(), customer.getPhoneNumber(), customer.getId()));
     }
 
     private boolean isUpdated(int updatedCount) {
